backend/requirements.txt
alembic==1.15.2
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Cors==5.0.0
Flask_JWT_Extended==4.7.1
flask_marshmallow==1.3.0
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
python-dotenv==1.1.0
SQLAlchemy==2.0.36
Werkzeug==3.1.3
psycopg2-binary==2.9.10
